stock-server/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-simulator-backend"
version = "0.1.0"
description = "A backend server for the Stock Simulator Demo"
authors = [{name = "Frodi Karlsson", email = "frodikarlsson@gmail.com"}]
dependencies = [
    # Flask for handling HTTP requests
    'flask >= 1.1.2',
    # Flask-CORS for handling Cross-Origin Resource Sharing
    'flask-cors >= 3.0.9',
    "python-dotenv >= 0.15.0",
    "pandas >= 1.1.3",
    "yfinance >= 0.1.55",
]
